Return 0/empty refund values in WC_Subscription
The WC_Subscription class needs to extend WC_Order instead of WC_Abstract_Order
to take advantage of the Edit Order screen, as this screen requires some refund
related methods to exist. However, a subscription doesn't currently have a refund
system, and it doesn't actually make sense for one either as a subscriptin order
is more a promise for future payments, not a record of historical payments.

These can be removed if WC core is patched to no longer require the refund API
methods and therefore, WC_Subscription can extend WC_Abstract_Order

Part of #5