Catch exceptions when updating sub status data
SHA: 24c7a4362b7b0 changed `WC_Subscription::update_status()` to update the `post_status` value
before updating related data (e.g. user's role, payment dates etc.), this helps to avoid issues
with methods which expect the subscription's status to be the new status, like
`wcs_maybe_make_user_inactive()` (#71), but it also creates the opportunity for error when
the data updates trigger an exception or do not complete correctly - the `post_status`
would be set to the new status, but the related data would be stuck at the values for
the old status.

By wrapping the actual logic to update the status into a `try/catch` block, we can
restore the old status if an exception occurs before passing the exception on for
the calling code to handle as needed.