Move removal of 'Pay' button to WCS_Cart_Renewal
Previously, `WC_Subscriptions_Checkout` was responsible for checking if
a manual or failed renewal order had been paid by some other method, and
therefore, the 'Pay' button shouldn't be displayed next to that button.

Goodness knows why it was in `WC_Subscriptions_Checkout`, but it's now
found a suitable home in `WCS_Cart_Renewal`.

Part of #32