Add wcs_create_resubscribe_order()
A wrapper for `wcs_create_renewal_order()` that sets a new piece of
meta data to record the original subscription for the order.

Also removes the `'_original_order'` post meta value from renewal
orders because this is no longer relevant in 2.0 - the subscription
stores the relevant data, not the original order.

Part of #32