Dont try save interval and period if not necessary
Needed also for #91.
When trying to save the subscription after I edited it, I noticed that it was trying to save non-editable fields (interval and period). When the next payment date cannot be changed, the interval and period fields also become plain text (instead of being drop down menus), therefore they won't be set in the POST data in some cases.