Do not allow customer to reactive pending sub
If something cancels the last renewal order for a subscription,
then Subscriptions would think it didn't require payment and display
the "Reactivate" button next to the subscription on the My Subscriptions
table. However, the subscription still needs payment and the customer
shouldn't be able to reactiate it until that order is paid (or the store
manager reactivates the subscription for them).