Stores download permissions on subscription as well
Part of #83

Mirrors code from the wc_downloadable_product_permissions function in WC core, and stores permissions against the subscription object as well.

Not sure about the placement of the hook / function. WC_Subscriptions class is always instantiated, whereas WC_Subscription only on demand, therefore adding the add_action to WC_subscription class runs the risk of not firing because WP would not know about it. The function itself is an integral part of the subscription, hence it's a static function on WC_Subscription.