Hide redundant or incompatible bulk actions
On the Subscriptions list table, do not display redundant actions,
e.g. "Put On Hold" if the list table is being filtered to display
only subscriptions that are on hold.

Also do not display incompatible actions. e.g. "Activate" if the
list table is being filtered to display only cancelled or expired
subscriptions (which can not be activated).

Part of #8