Use 'wcs_' prefix on user function hooks
To comply with new policy of using 'wcs_' on function hooks and the longer
'woocommerce_subscription_' prefix on class method hooks.

Also ensures that we use new hooks for v2.0 where new data is being passed
as parameters to callbacks on those hooks, which will be important with the
work in #66.