Pass the subscription to callbacks
The 'wcs_calculate_next_payment_from_last_payment' allows extensions to
determine if a subscription's next payment date should be calculated from
the date its most payment was due or the date its most recent payment was
actually processed (which in the case of manual renewals and failed payments,
means the two dates will differ).

However, it did not pass the WC_Subscription to which the calculation related
meaning it could only be done on a per site basis. There are good use cases for
filtering on a per subscription basis. For example, having all synchronised
subscriptions calculate from the date the payment was due to maintain the sync'd
schedule while leaving default behaviour for other subscriptions.

This patch passes the subscription to callbacks attached to the filter.