Use date_i18n() as it functions not as documented
`date_i18n()` is a tricky beast. Both the codex and the PHP Doc in the source code describing
its parameters say the 3rd parameter is a `boolean` to flag whether to "convert to GMT for time".

When reviewing the source of the function, it turns out it is used to determine whether to use
gmdate() or date(), but as WordPress always sets the PHP timezone to be UTC/GMT, there will be
no difference between the return values of those functions, because they are both using the same
timezone.

As a result, there were a few places in the source where I followed the docs thinking date_i18n()
would convert timestamps to the sites timezone, only to find they were actually in GMT/UTC still.