Don't use the next payment date when calculating if it's 0
Prior to this patch, when calculating the next payment date, if the next
payment was 0, it would be used as the base to count from but then it
would reach the maximum number of iterations in the while loop (30) and
exit. Meaning the new next payment method calculated would be 30 x period
from 1970-01-01.