Handle exceptions on date migration to v2.0
Because v2.0 validates dates using new rules, it's possible during the upgrade
that exceptions will be thrown because the dates were not required to conform
to these rules previously.

To work around this, this commit first attempts to massage the data in to a format
which conforms to the new rules and then it catches any other exceptions just in case
to make sure that it doesn't block the progress of the upgrader.

Part of #128