Allows 3rd party plugins to save meta on subscription object
Fixes #400

The reason for adding the original action hook next to the subscription one is that renewal orders are based on the subscriptions, not the original orders. That means for plugins that rely on setting order item metas on orders, 2.0 breaks compatibility.

Alternatively we can ask 3rd party developers to add support for subs 2.0. In the case of product addons, it would be one line, another add_action but for subs' hook.

The nested foreach is required to carry over multiple values of item meta.