Adds handling subscription download permissions on renewals as well
For #575,
Previously we only updated download permissions for a subscription when
dealing with with initial orders, now we want to also update and revoke
the permissions on renewal orders. This will mean that at the time of the
renewal, subscriptions will have up to date downloadable files without
the admin manually having to manage this.