Update update_status() param spec for WC 2.4.5
With woothemes/woocommerce@57bb0ebe WooCommerce introduced a new $manual
param on the WC_Abstract_Order::update_status() method. This causes a
strict standards notice to be thrown as the method declaration with
WC_Subscription::update_status() is incompatible. It also results in
inconsistent behaviour between manual (admin) status changes on orders
and subscriptions - the former would set the admin user's name as the
author of the order note, while the later would not, because the
$added_by_user parameter on add_order_note() would default to false.
This fixes that.