Copies over order meta even for keys that were auto-created
Fixes #616

The problem was that certain values for meta keys weren't being copied over. That was due to using add_post_meta with unique turned on. That worked well, becuase if the meta did exist, it wouldn't add a new one.

wc_create_order however populated a number of meta keys with empty values before we had a chance to get meta values over to the order from the subscription. That meant that due to the unique constraint those values never got copied.

The solution retains this ability, but also adds a way to update empty meta values, as the fourth parameter for update_post_meta is the previous value, which is empty by default. If the meta key doesn't exist, it falls back to add_post_meta.