Map core WC order status to WCS status
To avoid throwing exceptions when a plugin is working with a subscription the way
it would work with another WC_Abstract_Order descendent.

We still haven't mapped refunded (i.e. 'wc-refunded') because that is rarely
used and there is no equivalent for it on a subscription, so an exception probably
is warranted.