Correctly set order type on WC_Subscription object
With PHP 5.6, $subscription->order_type was being set to the parent's
value of 'simple' (probably as part of parent::__construct() call).

This was causing a fatal error in WC_Subscriptions_Synchroniser::maybe_add_subscription_meta()
when upgrading subscriptions containing a synchronised product, becasue
wcs_get_subscriptions() checks for $subscription->order_type, and it was
incorrectly set to 'simple', meaning no valid subscription would be returned
even though one existed.

This should be set to 'shop_subscription', so we're now setting is
both before (in the class properties) calling the parent's constructor
and afterwards.