Rename get_email_download_links()
To be get_item_downloads(), which more accurately represents its purpose
and conveys the hook it is attached too. As the 'woocommerce_get_item_downloads'
filter is called whenever WC_Abstract_Order::get_item_downloads() is called,
including when displaying the order on the My Account > View Order page,
not just when creating download links for emails.