Fix ability to resubscribe
Fixes #700

The problem arose because maybe_setup_cart() was using the request URI
to populate the cart. Because there was a resubscribe info set, it was
calling setup_cart, which adds products, which checks whether those are
purchasable.

That method was calling wcs_cart_contains_resubscribe, which was
returning false because the cart is not in session yet (ie WC()->cart
and WC()->session->cart) are both empty when it comes to products,
because no products have been added yet, because we're currently
checking whether the first one is purchasable.

The fix assumes that the product that we're allowing in the cart is the
one that we're trying to resubscribe to.