Change wcs_resubscribe_order_created to an action
Because the old parent renewal hook was an action, when I updated the Stripe and Simplify
Commerce code to handle removing order meta from resubscribe orders using the new
'wcs_resubscribe_order_created' hook, I treated it as an action, not a filter (i.e. I didn't
return the $resubscribe_order param for other methods to use).

Because of this, I think we need to use an action here instead of a filter. This is a real shame
because it would have been better to match this hook with the related 'wcs_renewal_order_created'
hook, but doing so would require patches to WooCommerce Core, Stripe and possibly other 3rd party
code.