Set subscription shipping based on recurring_total
Make sure we set the calculation type to 'recurring_total' when setting
the shipping line items on a subscription, because we need to call
WC_Cart->get_shipping_packages(), which will only filter recurring items
when that calculation type flag is set.

This fixes an issue with one time shipping where the initial order's shipping
line item was still being added to a subscription, even though it was not
being set in the total.

Part of #677