Make sure we update all package details
Instead of just removing the one time shipping products from the package,
make sure we also reduce the content cost of the package, in case there are
multiple items, some which do need shipping, some which do not, and the
shipping method uses the contents costs.

Also unset the package completely if the only item to be shipped was an item
with one time shipping.

Part of #677