Un-deprecate WC_Subscription_Order::get_recurring_total
Fixes #783

The deprecation notice says that instead of calling this function, we
should grab all the subscriptions and tally them up. Then proceeds to
grab all the subscriptions and tally them up. If we were to implement
the same behaviour, we'd arrive at the exact same function body that
this function currently has.