Fix Standard_IPN_Handler::get_order_id_and_key()
Reinstate the use of get_posts() as that is the simplest method for
finding a matching subscription and it saves all the version juggling
of IPN data. However, if we can't find a matching subscription, also
fix handling of JSON in the 'custom' param created with Subscriptions
prior to 2.0 (but with WC >= 2.3.11).