Create renewal order for PayPal failure
PayPal will send the recurring_payment_suspended_due_to_max_failed_payment
and subscr_failed IPN messages when a recurring payment has failed. But when
call $subscription->payment_failed(), the last order will be marked as failed
because it assumed prepare_renewal() has been run. This can mean either a order
for a previous successful renewal or even the original payment will be marked
as failed incorrectly. This patch fixes that by generating a renewal order before
calling $subscription->payment_failed() so that that renewal order will be marked
as failed and can be used by the custom to pay for the failed renewal and reactivate
the subscription.