Fixes infiniloop on upgrade
Fixes #799

Again down the rabbit hole of SQL, this two line fix took the better
half of the day.

So basically the original idea was to change the javascript that handles
counting whether it's updated or not. That wasn't sufficient, as Subs
still needed to call upgrade_complete, which it would only do if the
number of subs to upgrade left was 0. If we only change the javascript,
that doesn't affect the server at all, so we get into a different loop
(basically restarting the process, as active version never gets saved).

The underlying problem was that the order items have disappeared, but
the queries still returned results. For one of them it was one row with
a distinct order_item_id with a value of NULL. For the other, it was the
whole metadata, with the order_item_id being NULL. In the upgrade we
check whether we have an order_item_id, and if we don't, we skip it.

That meant that the upgraded subscription count never gets incremented
as that whole block is bypassed, but the number of rows are still not 0,
because there IS a result, it just happens to be NULL.

Hence this fix. :)