Don't always cancel subscriptions at PayPal
When a subscription expires or is pending cancellation, don't cancel it at
PayPal yet, instead just put it on hold. PayPal Standard is a one-way street,
once we cancel it, there's no going back. And if there is a bug setting an
incorrect expiration date (e.g. #824), that becomes a real problem.