Don't set a different trial end date
The upgrader bug would only set an incorrect trial end date if no end date
was set, that means we only need to check if the subscription now has an
end date and previously didn't, we don't need to check if the date is
different. This also reduces the possibility of modifying a trial end date
that a store manager has manually changed after the upgrade.