Repair recurring line tax data
Because in v2.0.0 and v2.0.1 it was not migrated correctly, as fixed
in #849.

We also want to run this repair even on cancelled subscriptions, so
we move that check after this. That ensures that resubscribes etc.
will have the correct base data and that the Edit Subscription
screen always shows correct taxes.