Display subscription details and notices
On the Order > Pay page when a changing payment method and the gateway
has redirected itself back to that page. This brings the page more
inline with what is display by WC core and makes sure any errors that
occur when loading the page are displayed to the customer rather than
a blank page.