Always use gmdate(), never date()
Because if some 3rd party code on the site is changing the server's
timezone with date_default_timezone_set(), using date() to simply
change the format of a timestamp can also change the actual value
of the timestamp. Specifically, date() will assume the timestamp
is in UTC and therefore, add or deduct any UTC offset for the
server's timezone before converting the timestamp to the chosen
format.

For example, all of the below output is for the same timestamp:

[UTC] gmdate( "Y-m-d H:i:s", $timestamp )                 = 2015-10-29 05:45:12
[UTC]   date( "Y-m-d H:i:s", $timestamp )                 = 2015-10-29 05:45:12
[Australia/Sydney] gmdate( "Y-m-d H:i:s", $timestamp )    = 2015-10-29 05:45:12
[Australia/Sydney]   date( "Y-m-d H:i:s", $timestamp )    = 2015-10-29 16:45:12
[America/Los_Angeles] gmdate( "Y-m-d H:i:s", $timestamp ) = 2015-10-29 05:45:12
[America/Los_Angeles]   date( "Y-m-d H:i:s", $timestamp ) = 2015-10-28 22:45:12