Fix logic checking for sync day in future
By comparing the full date not just the day then the month/year if
that day is after the current day. This ensures we don't see a day
in the future with a month and year in the past or same as today
as being a date in the future.

Fixes #959