Filter needs_payment when processing change payment method requests
I found the issue to be caused by both pay-form.php and
change-payment-method-form.php templates where both being loaded and
calling `$gateway->payment_method()`. Keep in mind the only way to
replicate this issue is to trigger a failed or pending-payment renewal,
manually mark the subscription as active and choose the change payment
method button on view-subscriptions. The solution i went with was to
simply filter needs payment on the subscription when processing a change
payment method request. This is to avoid the pay-form.php ever being
loaded but will make the `ob_clean();` usage redundant.