Use session token instead of current user id
If we use only the user ID, then we run into the problem where users are
sharing an account, as two different people could cause a temp table to
drop before it's being used in a query by another person.

Using session tokens is better, because each token depends on:
- the current user's id
- the user agent of the browser
- the remote address of the connection
- a 43 char random string

It's also BC, because it was introduced in WP 4.0, which WC 2.3.0
requires.