Delete subscriptions on user deletion
Don't just trash them. This makes sense now that a subscription is
distinct from an order and purely a promise for the future (so it
can be deleted without changing the record of history).

Part of #916

We could honour the $reassign param passed to callbacks on 'delete_user'
to reassign subscriptions to a different user, but I've decided not to as
I fear that type of automatic change may have a more negative impact that
simply deleting the subscription (i.e. it could lead to charging someone
incorrectly for a subscription, or a subscription being charged to one
customers credit card even though it is assigned to a different user in
the WooCommerce store). Deleting it is safer and the store manager can
manually change it before deleting the user if they want to.