Process initial payment via DoExpressCheckout
Previously, when purchasing subscriptions in an order with an initial payment
(either due to the subscriptions having sign-up fees and/or no free trials or
becuase other items are being purchased in the same order), we would simply
create a billing agreement and then immediately call DoReferenceTransaction
with that agreement to process the payment for that transaction. This helped
avoid having to support two equivalent API operations that PayPal uses different
API parameter keys for (i.e. DoReferenceTransaction and DoExpressCheckout).

The downside of this approach is that when arriving at PayPal, the customer is
only shown a very vague description of the billing agreement they are agreeing
to. By using DoExpressCheckout to process the initial payment, we can set the
current order's details in the SetExpressCheckout request, which means PayPal
will display those order details to the customer at PayPal (for an example of
the improved UI, see this screenshot: https://cloudup.com/cywEp0edQni ).

This commit changes things to use DoExpressCheckout for that reason.

Some gotchas worth discussing:
* although equivalent, the DoReferenceTransaction and DoExpressCheckout API
  endpoints require differently keyed parameters. Specifically, parameteres
  which are deprecated on DoExpressCheckout are the only available options
  for DoReferenceTransaction (specifically, the new parameters used by the
  DoExpressCheckout API have a PAYMENTINFO_0_ prefix). Because of this, we
  need to perform almost identical API calls with different parameter keys.
* similarly, the responses are also equivalent but have different parameter
  keys. To handle this, WCS_PayPal_Reference_Transaction_API_Response_Payment
  has been modified to handle responses to DoExpressCheckout API operations
  and WCS_PayPal_Reference_Transaction_API_Response_Recurring_Payment has
  been added to handle DoReferenceTransaction API operations. The reason
  for changing the existing class to handle the new operation instead of
  handling it is evident by looking at the new class - it simply uses no
  payment prefix.
* Because DoExpressCheckout operations also create a billing agreement, we
  also want to make WCS_PayPal_Reference_Transaction_API_Response_Payment
  extend the WCS_PayPal_Reference_Transaction_API_Response_Billing_Agreement
  class so that it has the get_billing_agreement_id() method.

Add:
* WCS_PayPal::process_subscription_payment_response()
* WCS_PayPal_Reference_Transaction_API::do_express_checkout()
* WCS_PayPal_Reference_Transaction_API_Request::do_express_checkout()
* WCS_PayPal_Reference_Transaction_API_Request::add_payment_details_parameters()
* WCS_PayPal_Reference_Transaction_API_Request::skip_line_items()
* WCS_PayPal_Reference_Transaction_API_Request::round()
* WCS_PayPal_Reference_Transaction_API_Response_Recurring_Payment