Apply existing renewal discounts as pseudo coupons
For coupons applied to the original subscription that still exist we can re-use the available information on the coupon.

For coupons that don't exist anymore we can apply a pseudo cart-based coupon to apply the discount.

If no coupons are applied but there is still a discount applied to the subscription (i.e. manually added discount) we also apply the discount using a cart-based coupon.

Sets products on the renewal order as product_ids - to help ensure coupons only apply to relevent products in cart.

Stores coupon info in a session variable so that the coupon info can be grabbed later and the discount applied.