Plop base taxes on top of line subtotal instead of adding line tax
Fixes #1097

The source of the problem is that whatever price calculate_totals would get, it would try to remove the base taxes from that number, if setting was included taxes.

In the item to renew, the line subtotal is always ex tax, as they are separate line item. The tax line items are the actual taxes as calculated by shipping address / billing address / shop base, but the point is, it's not always the same as taxes for the shop base.

All we have to do then is take the ex tax price, and add all the taxes that would apply at shop base. Because calculate totals would peel those off anyways, and then calculate the appropriate taxes as per customer's data.