Show recurring shipping selection fields if needed
It is possible for a store to offer shipping methods which are available for
the initial cart but not for a recurring cart (and/or vice versa). In these
cases, we need to display shipping selection fields for the recurring cart/s
so the customer can choose the appropriate shipping method for that cart.
Normally, we use the same method chosen on the initial cart, but that's not
possible in these cases.

This patch introduces the code to check whether a recurring cart's shipping
method keys match the initial cart's method keys. If they do not, it will
display shipping selection fields for all available rates for that cart.
If they do match, it will display the shipping amount for that cart based
on the shipping method chosen for the initial cart (the former behaviour).

Introduces:
* /tempates/cart/cart-recurring-shipping.php template for displaying
  the recurring shipping method selection field rows
* WC_Subscriptions_Cart::add_shipping_method_post_data() to make sure that
  the chosen recurring shipping method is available to save on checkout
* WC_Subscriptions_Cart::add_shipping_method_post_data() to make sure that
  the chosen recurring shipping method is saved when selected on checkout
* wcs_cart_totals_shipping_html() to decide what to display for each carts
  shipping row (i.e. either the total if the rates for that cart match the
  main cart, or selection fields if there are different rates)

Fixes #1178