Do not duplicate trial shipping methods
If the cart only contains a subscription product/s with free trials
or sync'd to a date in the future, the shipping packages will differ,
but the initial shipping is unnecessary, because nothing will be
shipped at the outset (all methods would be $0, but the selection
fields were still displayed).

Now that we have recurring shipping selection fields (with #1178)
we can use them instead of initial fields. We definitely don't want
to display both initial and recurring fields for the one set of
packages just because the rates differ, which we were doing, but
this fixes up.