Check product type when bulk editing variations...
...instead of whether it is a subscription product since 2.5.

Pre 2.5 we don't have the product type information available so we have to check if it is a subscription - downside here is this only works if the product has already been saved

Since 2.5 we have the product type information available so we don't have to wait for the product to be saved to check if it is a subscription.

Fixes the issue where bulk editing variations doesn't work until the product is saved.