Use chosen method not default method
To make sure the 'woocommerce_shipping_chosen_method' filter is called
for each recurring cart's shipping calculations, we filter the package
rates for that calculation becuase WooCommerce will apply the filter if:

1. there is no chosen method; or
2. the number of rates available for a given package differs to the session's
   method count; or
3. the chosen method is not in the available rates for the current package.

By filtering the package rates (first introduced with SHA: c92817f3c) we make
sure that the 3rd of those conditions is true for recurring cart calculations
where package rates differ.

However, as reported in https://github.com/Prospress/woocommerce-subscriptions/pull/1187#issuecomment-187996403
that also breaks shipping selection on the initial cart because it will also
cause the 2nd of those conditions to return true for the final shipping
calculation done to update packages for the initial cart (i.e. the 'none'
calculation) because the method count stored in the session is for the
previous recurring cart calculation, and it may have a different number of
methods than the current cart calculation.

This alone is fine, except whenever any of those conditions are met, WooCommerce
will use the default shipping method, instead of any chosen shipping method,
which means a customer can never choose anything other than the default shipping
method. WooCommerce 2.6 has fixed this issue by making sure that if a customer has
chosen a shipping method, WC_Shipping::get_default_method() will use that method
instead of the first method. This was perhaps an undiagnosed bug in WC 2.5 because
it would rarely ever be encountered and appears to have been unintentionally fixed in
https://github.com/woothemes/woocommerce/commit/9818796d3ef8d1dc691c90ac17601781460618d7

However, this also fixes that by making sure that the chosen shipping method is used
when it exists in the available rates with WC < 2.6.

It relies on the package index being passed to callbacks attached to the
'woocommerce_shipping_chosen_method' filter (which it currently is not), but that will
only cause issues when there is more than one package (i.e. index > 0) which is rare.