Modify both rate count & rate contents
When WooCommerce calculates rates for a recurring shipping package,
we want to return both a different number of rates, and a unique set
of rates for the recurring shipping package to make sure WooCommerce
updates the chosen method for the recurring cart (and the
'woocommerce_shipping_chosen_method' filter is called, which we use
to make sure the chosen method is the recurring method, not the
initial method).

Previously, SHA: c92817f3c315bc was introduced to make sure that the
number of rates was different, but the contents of the rates were not
so WC_Shipping::calculate_shipping_for_package() was still using the
cached copy of the rate, instead of a unique recurring amount.

SHA: ee3fc377c1022 was originally introduced to fix this issue, but
that needed to be fixed with SHA: 2ba7e7cbe55 instead.