Revert "Add view-subscription query var to WC's query vars array"
This reverts commit b785e8c74e9c838f8a764d3b615ef59da1568c76.

Although SHA: b785e8c7 fixed the issue outlined with #1088, it also
introduces a new, more sinister issue, reported in #1285.

The reason for that is because our endpoints do not get correctly
registered before WC_Install::install() calls flush_rewrite_rules().

This makes sure they are, though it also reinstroduces #1088 so an
alternative patch for that is coming up next.