Add WCS_Query::pre_get_posts()
The WC_Query::pre_get_posts() method says it is "to do the main product
query". However, it also does another extremely important task: fixing
endpoints on the front page. Because it only does this for its own query
vars (i.e. $this->query_vars), the fix was not being applied for the
'view-subscription' query var.

Initially, SHA: b785e8c7 was introduced to fix this. It added the
'view-subscription' query var to the WC()->query->query_vars so that
WC_Query::pre_get_posts() would fix the issue for 'view-subscription'
requests; however, that also introduced #1285, so it was reverted with
SHA: 08bc9efc91a6.

This patch instead takes the parts of WC_Query::pre_get_posts() we need
to fix our own query vars on the front end and makes sure they are applied
for our query vars (i.e. 'view-subscription').