Pass the recurring package key as index value
Pass the entire recurring package key as the $index value to the
cart-recurring-shipping.php template instead of just the $i value.

This fixes compatibility with Local Pickup Plus, which uses the $index
value to link the pickup location with package selection fields.

This hopefully also brings the use of cart-recurring-shipping.php a bit
closer to the WC behaviour, by having the $index value be the actual value
used as the key on the fields, instead of only part of the key. If the
cart-shipping.php is eventually changed to accept string indexes instead
of just integer indexes, we should be able to use the same code as we have
now with cart-shipping.php.

Partial fix for #1289