Recalculate all dates for sync'd products
If payment is processed on an order more than an hour after the order
and subscription were initially created, we update the dates on the
subscription to account for a new start date of the time the order's
status is changed (rather than the time it was created).

However, if the subscription has a sync'd date, we previously only
recalculated its next payment date, end we only did that if the
existing next payment date was before the time the order's status
is changed. However, we also need to recalculate the trial end date
to make sure we don't have trial end date that occurs on the same
day as the first renewal date, but at a time after 3:00am (see #1328
for more). And we also need to update the end date to make sure that
it is based on the next next payment date.

This patch makes sure all 3 dates are recalculated whenever either the
trial end or next payment date are set on a subscription and those dates
occur before the time the order's status is changed (i.e. payment is
processed).

Part of the fix for #1328 and woothemes/woocommerce-subscriptions#403

Related to issues on #1324