Add recalculate_product_trial_expiration_date()
To make sure that if the trial end date occurs on the same day as the
first renewal payment, but at a later time, we set the trial end time
to be the same as the first renewal time (i.e. 3:00am site time) to
avoid error messages being displayed on checkout.

Part of the fix for #1328 and woothemes/woocommerce-subscriptions#403

Related to issues on #1324