Take control of Related Orders meta box CSS
WooCommerce 2.6 changes the order items CSS, breaking the display
of the Subscriptions related order items table, which piggy-backed
off the order items CSS.

To fix this, we should take control of the CSS used to display that
metabox. By doing so, we can also remove the order item classes and
hardcoded CSS from the metabox HTML.

Part of #1338

This also tweaks the display a little to remove extra whitespace from
separate table borders and padding.