Backup shipping methods only once per calculation
SHA: c1bb25ccec8 introduced a special 'ost_shipping_methods' session
value to backup shipping methods to make sure we could restore the
correct methods if WC or Subscriptions wiped out shipping methods to
account for a product which required one-time shipping only. This was
important to make sure that when a shipping method was chosen on the
cart page, that selection persisted to the checkout page (as reported
in #749 and address in #863).

However, once #1187 was introduced, it also introduced the concept of
recurring shipping methods (i.e. a shipping method applied only to a
recurring cart, not the initial cart). Along with this came chosen
shipping methods using a recurring cart ID.

With these two combined, we can come across an interesting scenario:
adding a product to the cart which requires selecting a recurring
shipping method (because its packages differ to the initial package,
probably because of different shipping classes) and then removing
that product and adding a product with an identical billing schedule
but that does not require recurring shipping would mean the recurring
shipping method selected for the other product would persist to the
new product, regardless of what shipping method was chosen for the
new products initial shipping method because the old product's
recurring method was cached in 'ost_shipping_methods' with an identical
recurring cart key/package index to the new method.

This could be fixed by forcing the use of the initial package for
recurring shipping, but the 'ost_shipping_methods' cache would still
contain invalid data, we really only want that cache to be used
within the recurring cart calculations, not between requests, so
this fixes that.