Use renewal order total for scheduled payments
Instead of passing the subscription total to gateways as the amount that
should be charged for the renewal, use the renewal order's total, so that
extensions can modify the renewal order (via the `wcs_renewal_order_created`
hook or similar) and set a different amount to be charged.

Handy for enabling things like one-off discounts on the next renewal etc.
without requiring the subscription line total to be changed.