Assorted tweaks to clean up check for customer ID
Change function name to try to be more descriptive and conform
to other functions.

No need to log order note, just throw exception: order notes are
more useful for things happening in the background, or that others
may need to keep track of, not actions involving the store manager
which have no longer term impact that needs to be recorded.

Also add missing . to exception message/notice and explain that
it can be activated by adding a customer.

Also no need to have a return value sometimes but not others and
try to make whitespace compliant with rest of codebase.