Fix bbPress page title conflict
As Subscriptions was changing the page title on any page with the
query var "subscriptions" it was conflicting with bbPress, which has
pages for viewing forum subscriptions, and changing the title on
those pages.

This fixes that by making sure we only change the page title on
an account page, which we can trust will be a WooCommerce page.