Deprecate filter_package_rates()
Because we will use a different approach to modify the number of
available rates and ensure chosen methods are honoured for recurring
totals (by changing the value set in the session).

Although filter_package_rates() also claims to have been avoiding
WC's rate caching by setting the recurring calculationg type on each
rate (as introduced with SHA: 888cae5), this doesn't appear to be the
case because WC determine's a packages hash *before* apply the
'woocommerce_package_rates' filter, and it does not update the hash
after that filter, which means only the details of the package before
the rates are calculated/set are used to determine the hash.

The package details are being modified already to circumvent the cache
with set_cart_shipping_packages() which is setting a recurring_cart_key
value on each package.

Changing this apprach also changes the need for the horrible hack of
'wcs_dummy_rate', which is up there with the grossest hacks I've ever
written.

This is part of #1583