Add reset_shipping_method_counts()
To temporarily reset the number of shipping methods stored in the
session and make sure that 'woocommerce_shipping_chosen_method' is
triggered by WooCommerce so we can set the correct recurring methods.

Previously, this was handled in filter_package_rates() by reducing the
rate count only to the chosen recurring rate. Once caching of calculations
was added to Subscriptions with #1485, this introduced a bug because
rates weren't recalculated again when displaying the available methods,
and the cached rates did not contain all the shipping methods for that
package (it only contained the rate for the *chosen* method, as that's
the only rate we included), as reported in #1583.

This alternative approach allows WooCommerce to cache all the recurring
rates for each recurring package, while ensuring that we can set the
chosen shipping method correctly (via woocommerce_shipping_chosen_method).

The reason it was not chosen initially is because it is modifying core
data and introducing the possibility of a plugin conflict. However,
becuase we hook to 'woocommerce_shipping_packages' with a very high
priority, change of conflicts with 3rd party plugins that may use the
method count session value is on callbacks attached to either
'woocommerce_shipping_chosen_method' or 'woocommerce_shipping_method_chosen',
as the session value is updated again by WC_Shipping->calculate_shipping()
soon after we modify it.

Fixes #1583